Added
Added a method to be able to set the limit for a strip.

Added a vban show chat method to the command class,

Added return methods to the command class methods.

Modified the macro button class to default to the state of the button rather than nothing.
eg. macroButton.getstatus(3) will not error anymore.